added testing facilities to selection-test // finally fixed selection
but I don't know why this 'fix' works:
The rendered image is the distorted one, not the image being selected
from.  The only difference is that during selection, the projection
matrix is set up freshly, i. e. load-identity set-perspective.  If you
do this for rendering, the distortion disappears.  But the projection
matrices before and after this reproduction are IDENTICAL!

It works now in selection-test anyway... but must be abstracted to
apply for all modules.  Could be integrated into non-agar-rendering,
instead of loading the saved matrix (which should actually be the same
but apparently, 1 != 1 ...) Or I just forgot something.