added other players' cards and bend cards, changed coordinate system
bending here means rotating the cards a bit around the y-axis and
placing them on a virtual circle in the x-z-plane

as for the coordinate system: the origin of the new coordinate system
is the middle of the table (on table level), i. e. cards are all moved
in positive z-direction (y-rotation into player perspective asserted)

fixed bugs for the other players' cards:
back-textures were not drawn because if you also supply the texture
parameter that would have overwritten the texture binding